Add/Change-duration-type-from-Long-to-float-in-Test-and-TestDTO

+++ src/main/java/com/abouals3ood/quizzify_server/dto/TestDTO.java
@@ -11,5 +11,5 @@ public class TestDTO {
 
     private String description;
 
-    private Long duration;
+    private float duration;
 }
